Refactor middleware and update logging syntax
Updated the way how middleware is being applied in the bot.py file. Instead of invoking middleware methods individually, instituted a for loop to wrap through all global middleware components. This streamlines the code and simplifies the introduction of new middleware.

In user.py, unnecessary middleware declaration for CheckUser() is also removed to avoid redundancy, as it's already included in the global middleware in bot.py.