"Remove unused migration files and update Docker and Database configs
This commit removes several unused migration files from the database's migration versions. It also introduces changes to the Dockerfile and docker-compose.yml files to accommodate a webhook service with appropriate ports and environment variables. The .env file was also updated to include necessary environment variables. New scripts for running and creating migrations were added in the 'scripts' directory. The NowPayments callback url was correctly updated. There are also several changes in the application logic, particularly related to payments processing, including usage of the 'NowPaymentsAPI' to create payment data and send messages to the 'bot' service when payments are received.

These changes were made to improve the handling of payments processing, cleanup unused database migrations, streamline the Docker configuration and improve the overall structure of the application."