Reorder middleware and fix user data retrieval in translator
This commit addresses two issues in the code. Firstly, the middleware order is adjusted in bot.py to ensure CheckUser middleware precedes TranslationMiddleware. This is to confirm user's identity before invoking the translation middleware.

Secondly, the user's data is now properly retrieved from 'user_info' instead of 'user' as previously, in the translator middleware. Also, the user's language attribute is now correctly accessed for the translator middleware. This forms the basis for accurate translation based on user's preferred language.