Change Docker port and update python version and invoice message
Updated the port number in the docker-compose.yml file from 8443 to 8001 for better port management efficiency. Python version in Dockerfile is changed from 3.10 to 3.11-slim for the availability of the latest features and security patches. In tg_bot/dialogs/selected.py, modified the invoice message to show the amount in coins instead of USD, to provide more user-friendly information as per new business requirements.