Update minimum payment and bot responses
Updated the minimum deposition and payment threshold in the telegram bot from 100 to 35 and 10 respectively. The Ukrainian and Russian locale files were updated to reflect these changes. An event handling for unrecognized user messages was added in the user router section. The NowPayments method was uncommented to allow users to choose it as a payment method again. The balance query in the base repo was updated to consider only successful transactions. Default values were set to None for several fields in the nowpayments types to avoid potential errors with uninitialized fields.