Correct get_balance method to only retrieve USD amount
Adjusted the get_balance method in getters.py to only fetch the user's balance in USD, instead of previously fetching both coins and USD. This change is in line with the recent alterations in the application, where we have moved towards dealing with only USD amounts. Therefore, the coin balance retrieval is redundant and has been removed for cleaner, simpler code. This also assures more accurate financial operations by using USD as the standard currency measure.