Refactor code to handle amounts in USD not coins
This commit changes the way amounts are handled in the Telegram bot. Previously, amounts were handled in terms of "coins", which were then converted to USD. To simplify the code and avoid confusion or errors due to conversions, we are now handling amounts directly in USD. All references to "coins" have been removed or changed to work with USD amounts. Also, the minimum deposit amount has been reduced from 10 to 7. Logging has been introduced to keep track of the operations.