Add usd_amount parameter to order handling methods
Updated the 'add_order', 'get_order_info', and 'transaction_minus' methods in repo/base.py to handle usd_amount. This change will allow these functions to work directly with the financial amount tied to an order instead of having to convert to and from "coins". This simplifies the operation, reduces the chances of error and makes the code more consistent in using the real world currency.