Update pre_confirmation_text logic to include usd_amount
Modified the pre_confirm_text getter method in tg_bot/dialogs/getters.py to include usdt_amount by inventing a new parameter in the the method. This change is meant to show monetary amount tied to the given url count. This should improve user understanding of the estimated costs associated with each url.