Add referral percentage constant to `misc/constants.py`.
This change defines a new `PERCENT` constant in `misc/constants.py` that represents the referral percentage. Earlier, the percentage rate (0.08) was hardcoded in multiple places. This update not only makes the code more maintainable and readable but also makes it possible to change the percentage rate in one place if necessary in the future.