Lowered link count requirement and added referral system function.
Link count requirement lowered from 10 down to 1 to allow functionalities with fewer inputs. Implemented function 'referral_system' which counts the number of referrals from a user and the total referral amount, then sends a message to the user regarding these numbers. This was added to keep track of user referrals and incentivize users to refer more.