Add ForeignKey to users model
Introduced ForeignKey into the users model to track referral data for each user. The 'referrer_id' attribute will link back to the initial user who referred them, enabling us to keep track of the referral chain. This is an important addition for the development of future features such as user referral incentives.