"Refactor referrer_id datatype in User model"
The datatype for the referrer_id field in the User model was modified from compulsory integer to optional integer. This was to account for new users who would not have a referrer_id initially. The referrer_id is now mapped under "Mapped[Optional[int]]" to allow for null values, enhancing support for new users without referrals.